Shift additional files fixtures location for reuse on more tests
